test(sync): move M0 sample contracts into reporting suite
